\name{BoostMLR.news}
\alias{BoostMLR.news}
\title{Show the NEWS file}
\description{
  Show the NEWS file of the BoostMLR package.
}
\usage{
BoostMLR.news(...)
}
\arguments{
  \item{...}{Further arguments passed to or from other methods.}
}
\value{
  None.
}
\author{
  Amol Pande and Hemant Ishwaran
}
\keyword{documentation}
